moved gc code into own files
